Updated Dashboard for Search
