+++ bookhub/src/App.css
@@ -19,6 +19,7 @@
   padding: 0;
   margin: 0;
   font-family: sans-serif;
+  overflow-y: scroll;
 }
 
 .search-bar {
@@ -44,7 +45,9 @@ input[type="text"] {
   background-color: var(--black);
   color: white;
   padding: 20px;
-  height: 125%;
+  height: 100vh;
+  /* justify-items: flex-start; */
+  /* display: flex; */
 }
 
 body {
@@ -90,19 +93,20 @@ body {
   list-style: none;
   padding: 0;
   margin: 0;
-  display: flex;
-  flex-direction: column;
+  /* display: flex; */
+  /* flex-direction: column; */
 }
 
-.options li {
+.options-link {
   margin-bottom: 10px;
   padding: 10px;
   transition: background-color 0.3s ease;
   font-size: large;
   font-weight: bold;
+  color: white;
 }
 
-.options li:hover {
+.options-link:hover {
   background-color: var(--gray);
   cursor: pointer;
 }
@@ -110,14 +114,19 @@ body {
 .search-bar {
   display: flex;
   align-items: center;
-  justify-content: flex-end;
+  justify-content: flex-start;
   margin-bottom: 20px;
   padding: 10px;
   background-color: var(--black);
+  font-size: larger;
+  color: white;
+  font-weight: bold;
 }
 
 .content {
   flex-grow: 1;
+  /* overflow-x: visible; */
+  overflow-y: scroll;
 }
 
 body {
@@ -179,8 +188,6 @@ body {
   left: 50%;
   transform: translate(-50%, -50%);
   padding: 2%;
-  /* display: flex; */
-  /* align-items: center; */
 }
 
 .display-block {
@@ -193,4 +200,15 @@ body {
 .modal-img {
 }
 .modal-details {
+}
+.navbar-expand .navbar-nav {
+  flex-direction: column;
+  justify-content: flex-start;
+}
+.sidebar-container {
+  margin-top: 0;
+  margin-bottom: auto;
+}
+.navbar-light .navbar-nav .nav-link {
+  color: white;
 }
