+++ bookhub/src/scenes/statistics/index.js
@@ -1,9 +1,12 @@
+import Topbar from "../global/Topbar";
+
 const Statistics = () => {
   // const [subjects, setSubjects] = useState([]);
   // setSubjects(["humor", "fantasy", "love", "magic", "romance"]);
 
   return (
     <div>
+      <Topbar title={"Books Statistics"} />
       <h1>Statistics</h1>
     </div>
   );
